Add preserve-indent command line directive
Use /PI to tell JCP to preserve line indentation when processing
line-prefixes //$ and //$$.  When /PI is specified, then "//$" gets
replaced with three spaces, and "//$$" gets replaced with four spaces.

More generally, if both /PI and /ES are specified, then a prefix
string of length n is replaced with a string composed of n
spaces. E.g. "//  $" is replaced with five spaces, and "//   $$"
is replaced with seven spaces.